Divide drop target component
