+++ website/frontend/src/components/pages/Programs/children/Upload/Upload.module.scss
@@ -1,0 +1,7 @@
+.upload {
+  display: flex;
+
+  & > *:not(:last-child) {
+    margin-right: 30px;
+  }
+}
